include protocol separator
B data is being written into what would is actually the "name" and "kwd..." metanet fields.

If there is no separator, there is no programmatic way to parse the data based on the protocol schema alone. You can't tell where meta stops and B begins. Looking for the specific B prefix string is not a solution as that would also be a valid filename or keyword string.

Many apps have been using the | character as a protocol separator.